feat(Blox.CommentController): use plug to find post
If the post is found then it is stored in the `assign` map, otherwise a
flash message is added and the user is redirected.

The `post_id` param is no longer pattern matched on because it is not
required.

The `Blox.ControllerTestCase` module has been modified to use
`Phoenix.ConnTest` instead of `Plug.Test` so that the phoenix helpers
can be used.